Repair AppImage Build
Ok so: don't take my word for it. Try it, it should work. Many users of electron-builder have reported that the AppImage build was not working. The workaround for now is 1. Moving back to Electron Builder 20.14.7 and 2. Providing a "files" entry in the package.json so that electron-builder includes those files in the AppImage build.