Add files via upload
I added a new button to activate the dark mode. It is still not completely working and has many issues. It is able to change the background of the emoji div. I am still working on it and would like some help.
I am a newbie to electron and have a question. After disabling and enabling a CSS file dynamically, how to restart/reconstruct the app?
If someone knows how to do that then just by making two CSS files with only background color changes we a dark mode template for the app very easily.